# packages/nvl/auth/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NVL Auth HTTP API",
    "version": "0.2.0",
    "description": "All operations are optional and controlled by package, surface, feature, and feature-route flags."
  },
  "servers": [{"url": "/api/v1/auth"}],
  "paths": {
...
    "/password/forgot": {"post": {"operationId": "nvl.auth.public.password.request", "x-nvl-feature": "password", "x-nvl-feature-operation": "issue", "x-nvl-route-surface": "public", "x-nvl-feature-dependencies": ["authentication"], "requestBody": {"$ref": "#/components/requestBodies/Identifier"}, "responses": {"202": {"$ref": "#/components/responses/Accepted"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}}},
...
    "/magic-links": {"post": {"operationId": "nvl.auth.public.magic_links.request", "x-nvl-feature": "magic_links", "x-nvl-feature-operation": "issue", "x-nvl-route-surface": "public", "x-nvl-feature-dependencies": ["authentication", "sessions"], "requestBody": {"$ref": "#/components/requestBodies/Recipient"}, "responses": {"202": {"$ref": "#/components/responses/Accepted"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}}},
...
    "/security-codes": {"post": {"operationId": "nvl.auth.public.security_codes.request", "x-nvl-feature": "security_codes", "x-nvl-feature-operation": "issue", "x-nvl-route-surface": "public", "x-nvl-feature-dependencies": ["authentication"], "requestBody": {"$ref": "#/components/requestBodies/SecurityCodeRequest"}, "responses": {"202": {"$ref": "#/components/responses/Accepted"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}}},
...
    "/social-identities/{socialIdentity}": {
      "parameters": [{"$ref": "#/components/parameters/socialIdentity"}],
      "delete": {"operationId": "nvl.auth.account.social.destroy", "x-nvl-feature": "social_identities", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "account", "x-nvl-feature-dependencies": ["authentication"], "responses": {"200": {"$ref": "#/components/responses/Success"}}}
    },
...
    "/api-tokens/{tokenId}": {
      "parameters": [{"$ref": "#/components/parameters/tokenId"}],
      "put": {"operationId": "nvl.auth.account.api_tokens.update", "x-nvl-feature": "api_tokens", "x-nvl-feature-operation": "update", "x-nvl-route-surface": "account", "x-nvl-feature-dependencies": ["authentication"], "requestBody": {"$ref": "#/components/requestBodies/ApiToken"}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}},
      "delete": {"operationId": "nvl.auth.account.api_tokens.destroy", "x-nvl-feature": "api_tokens", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "account", "x-nvl-feature-dependencies": ["authentication"], "responses": {"200": {"$ref": "#/components/responses/Success"}}}
    },
...
    "/invitations/{invitation}": {
      "parameters": [{"$ref": "#/components/parameters/invitation"}],
      "delete": {"operationId": "nvl.auth.management.invitations.destroy", "x-nvl-feature": "invitations", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}}
    },
...
    "/clients/{client}": {
      "parameters": [{"$ref": "#/components/parameters/client"}],
      "get": {"operationId": "nvl.auth.management.clients.show", "x-nvl-feature": "clients", "x-nvl-feature-operation": "read", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": ["authentication"], "responses": {"200": {"$ref": "#/components/responses/Success"}}},
      "put": {"operationId": "nvl.auth.management.clients.update", "x-nvl-feature": "clients", "x-nvl-feature-operation": "update", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": ["authentication"], "requestBody": {"$ref": "#/components/requestBodies/ClientMutation"}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}},
      "delete": {"operationId": "nvl.auth.management.clients.destroy", "x-nvl-feature": "clients", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": ["authentication"], "responses": {"200": {"$ref": "#/components/responses/Success"}}}
    },
...
    "/users/{user}": {
      "parameters": [{"$ref": "#/components/parameters/user"}],
      "get": {"operationId": "nvl.auth.management.users.show", "x-nvl-feature": "principal_management", "x-nvl-feature-operation": "read", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}},
      "put": {"operationId": "nvl.auth.management.users.update", "x-nvl-feature": "principal_management", "x-nvl-feature-operation": "update", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "requestBody": {"$ref": "#/components/requestBodies/UserMutation"}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}},
      "delete": {"operationId": "nvl.auth.management.users.destroy", "x-nvl-feature": "principal_management", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}}
    },
...
    "/users/{user}/permissions": {"parameters": [{"$ref": "#/components/parameters/user"}], "put": {"operationId": "nvl.auth.management.users.permissions", "x-nvl-feature": "principal_management", "x-nvl-feature-operation": "update", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "requestBody": {"$ref": "#/components/requestBodies/PermissionAssignment"}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}}},
    "/users/{user}/restore": {"parameters": [{"$ref": "#/components/parameters/user"}], "post": {"operationId": "nvl.auth.management.users.restore", "x-nvl-feature": "principal_management", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}}},
...
    "/roles/{role}": {
      "parameters": [{"$ref": "#/components/parameters/role"}],
      "get": {"operationId": "nvl.auth.management.roles.show", "x-nvl-feature": "rbac", "x-nvl-feature-operation": "read", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}},
      "put": {"operationId": "nvl.auth.management.roles.update", "x-nvl-feature": "rbac", "x-nvl-feature-operation": "update", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "requestBody": {"$ref": "#/components/requestBodies/RoleMutation"}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}},
      "delete": {"operationId": "nvl.auth.management.roles.destroy", "x-nvl-feature": "rbac", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}}
    },
...
    "/permissions/{permission}": {
      "parameters": [{"$ref": "#/components/parameters/permission"}],
      "get": {"operationId": "nvl.auth.management.permissions.show", "x-nvl-feature": "rbac", "x-nvl-feature-operation": "read", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}},
      "put": {"operationId": "nvl.auth.management.permissions.update", "x-nvl-feature": "rbac", "x-nvl-feature-operation": "update", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "requestBody": {"$ref": "#/components/requestBodies/PermissionMutation"}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "422": {"$ref": "#/components/responses/ValidationFailure"}}},
      "delete": {"operationId": "nvl.auth.management.permissions.destroy", "x-nvl-feature": "rbac", "x-nvl-feature-operation": "revoke", "x-nvl-route-surface": "management", "x-nvl-feature-dependencies": [], "responses": {"200": {"$ref": "#/components/responses/Success"}}}
    },
...
  },
  "components": {
    "parameters": {
...
      "socialIdentity": {"name": "socialIdentity", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "tokenId": {"name": "tokenId", "in": "path", "required": true, "schema": {"type": "string"}},
      "invitation": {"name": "invitation", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "client": {"name": "client", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "user": {"name": "user", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "role": {"name": "role", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "permission": {"name": "permission", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
    },
    "securitySchemes": {
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "laravel_session", "description": "Configured Laravel stateful guard cookie; the host may rename it."},
      "bearerToken": {"type": "http", "scheme": "bearer", "description": "Host account/management middleware may accept Sanctum or another bearer-token authority."}
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "required": ["data", "code", "message"],
        "properties": {
          "data": {},
          "code": {"type": "string"},
          "message": {"type": "string"}
        }
      },
      "LaravelValidationError": {
        "type": "object",
        "required": ["message", "errors"],
        "properties": {
          "message": {"type": "string"},
          "errors": {"type": "object", "additionalProperties": {"type": "array", "items": {"type": "string"}}}
        }
      },
...
      "Identifier": {"type": "object", "required": ["identifier"], "properties": {"identifier": {"type": "string", "maxLength": 255}}},
      "Recipient": {"type": "object", "required": ["recipient"], "properties": {"recipient": {"type": "string", "maxLength": 320}}},
...
      "SecurityCodeRequest": {"type": "object", "required": ["recipient", "purpose"], "properties": {"recipient": {"type": "string", "maxLength": 320}, "purpose": {"type": "string", "maxLength": 120}}},
...
      "ApiToken": {"type": "object", "required": ["name", "abilities"], "properties": {"name": {"type": "string", "maxLength": 120}, "abilities": {"type": "array", "minItems": 1, "maxItems": 100, "uniqueItems": true, "items": {"type": "string", "maxLength": 120}}, "expires_at": {"type": ["string", "null"], "format": "date-time"}}},
...
      "ClientMutation": {"type": "object", "required": ["name", "surface", "base_url"], "properties": {"name": {"type": "string", "maxLength": 120}, "surface": {"type": "string", "maxLength": 40}, "base_url": {"type": "string", "format": "uri", "maxLength": 2048}, "return_paths": {"type": "array", "items": {"type": "string", "maxLength": 2048}}, "allowed_origins": {"type": "array", "items": {"type": "string", "format": "uri", "maxLength": 2048}}, "allowed_flows": {"type": "array", "items": {"type": "string", "maxLength": 80}}, "metadata": {"type": "object", "additionalProperties": true}, "active": {"type": "boolean", "default": true}}},
...
      "UserMutation": {"type": "object", "properties": {"name": {"type": "string", "maxLength": 160}, "email": {"type": "string", "format": "email", "maxLength": 254}, "password": {"type": ["string", "null"], "minLength": 12, "writeOnly": true}, "active": {"type": "boolean"}, "email_verified": {"type": "boolean"}, "locale": {"type": "string", "maxLength": 12}, "timezone": {"type": "string", "maxLength": 64}, "profile": {"type": "object", "additionalProperties": true}, "preferences": {"type": "object", "additionalProperties": true}, "roles": {"type": "array", "items": {"type": "string"}}, "permissions": {"type": "array", "items": {"type": "string"}}}},
...
      "PermissionAssignment": {"type": "object", "required": ["permissions"], "properties": {"permissions": {"type": "array", "uniqueItems": true, "items": {"type": "string"}}}},
      "RoleMutation": {"type": "object", "required": ["name"], "properties": {"name": {"type": "string", "maxLength": 160}, "display_name": {"type": ["string", "null"], "maxLength": 160}, "description": {"type": ["string", "null"], "maxLength": 2000}, "parent_id": {"type": ["string", "null"], "format": "uuid"}, "priority": {"type": "integer"}, "permissions": {"type": "array", "uniqueItems": true, "items": {"type": "string"}}, "metadata": {"type": "object", "additionalProperties": true}}},
...
      "PermissionMutation": {"type": "object", "required": ["name"], "properties": {"name": {"type": "string", "maxLength": 160}, "display_name": {"type": ["string", "null"], "maxLength": 160}, "description": {"type": ["string", "null"], "maxLength": 2000}, "group": {"type": ["string", "null"], "maxLength": 120}, "metadata": {"type": "object", "additionalProperties": true}}}
    },
    "requestBodies": {
...
      "Identifier": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Identifier"}}}},
      "Recipient": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recipient"}}}},
...
      "SecurityCodeRequest": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SecurityCodeRequest"}}}},
...
      "ApiToken": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApiToken"}}}},
...
      "ClientMutation": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClientMutation"}}}},
...
      ,"UserMutation": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserMutation"}}}}
...
      ,"PermissionAssignment": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PermissionAssignment"}}}}
      ,"RoleMutation": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoleMutation"}}}}
...
      ,"PermissionMutation": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PermissionMutation"}}}}
    },
    "responses": {
      "Success": {"description": "Successful Auth operation", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store, private"}}, "Referrer-Policy": {"schema": {"type": "string", "const": "no-referrer"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}},
...
      "Accepted": {"description": "Auth request accepted without account disclosure", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store, private"}}, "Referrer-Policy": {"schema": {"type": "string", "const": "no-referrer"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}},
      "ValidationFailure": {"description": "Laravel validation failure or neutral credential rejection", "content": {"application/json": {"schema": {"oneOf": [{"$ref": "#/components/schemas/LaravelValidationError"}, {"$ref": "#/components/schemas/Envelope"}]}}}}
    }
  }
}